requirements.txt
streamlit==1.28.0
google-genai==0.8.3
python-dotenv==1.0.0
